The pushd script as discovered
This commit is basically the script shown in the presentation. There
only difference is the addition of a Logger to demonstrate when a POST
is being made to the GCM API.  This is because the included Authorization
token isn't valid and staging a test like the one in the presentation is
not practical.

To test this script:

1. Run pushd

  bundle exec ruby ./pushd.rb

2. Use netcat to test PING:

  echo "PING" | nc -u 127.0.0.1 6889

3. Use netcat to test SEND:

  echo 'SEND abc123 "Hello, World!"' | nc -u 127.0.0.1 6889