Add AsyncHelper to clean up spec
We were using calls to sleep as a way to force thread synchronization.
This introduced two problems:

1. The tests were slow -- we were paying a tax to the sleep calls each
   time even though the tests might not have needed to wait at all.

2. Under certain circumstances (busy system, other workstations, CI),
   the tests sometimes failed spuriously.

Inspired by the GOOS book, the eventually helper assists in testing
asynchronous behavior by retrying a given expectation within a timeout
block. As we can't control when the scheduler will yield control back
to the thread running the behavior we're testing, we'll probe it until
it succeeds. if the test doesn't pass within a second, it'll timeout
and return the original exception.