Replace case statement with hash
Now that our case's branches are identical, we can put the last nail in
its coffin and eliminate it entirely. This is a good thing because it prevents
any custom command-specific logic from slipping in the branches. Instead of
using a conditional, we're using a hash manifest that maps a command to a
specific job class. Our factory uses the incoming command to select the
right class from the hash and if it finds one, instantiates and returns it.