Wrap sockaddr structure in Client object
We've been passing around a parameter called data which is terrible
for a number of reasons: data doesn't mean anything and its contents
are actually quite important as they are the incoming message and
remote client details.

The latter -- the remote client details -- are passed in as an array
referred to as the sockaddr structure. This change introduces an object
to encapsulate sockaddr's contents. It uses the Addrinfo object from
the standard library which knows how to parse this structure and
decorate it with friendlier methods.