Add send method to Client
The Ping job was plucking data from the Client in order to send a
response back to the Client through the server. By adding a send
method onto the Client we can tell the Client to send a message
and have it collaborate with the server and its data.

As a result we can stop passing server around as we're going to
encapsulate the server within the Client.