Introduce PushNotification object
Since we have two concepts of 'client' in the Send job, let's extract
one of them to its own object. The Send job has a bunch of code specific
to the Google Cloud Messaging API so moving those concerns out focuses
the job on parameter extraction and instantiation of a new push notification.